Fix(Color picker) import service
