+++ src/demo-app/colorpicker/colorpicker-demo.html
@@ -1,9 +1,14 @@
 <div class="demo-colorpicker">
-  <h1>Colorpicker Demo</h1>
-  <hr />
-  <h4>Color Picker</h4>  
-  <div>    
-      <md2-colorpicker [(ngModel)]="color" (change)="change($event)">
-      </md2-colorpicker> 
-  </div>
+    <h1>Colorpicker Demo</h1>
+    <hr />
+    <h4>Color Picker</h4>  
+    <div>
+        <md2-colorpicker [(ngModel)]="color" (change)="change($event)">
+        </md2-colorpicker> 
+    </div>
+    <h4>Disabled</h4>
+    <div>
+        <md2-colorpicker [(ngModel)]="color2" (change)="change($event)" [disabled]="true">
+        </md2-colorpicker> 
+    </div>
 </div>
