+++ src/lib/index.ts
@@ -5,7 +5,7 @@ export * from './accordion/index';
 export * from './autocomplete/index';
 export * from './chips/index';
 export * from './collapse/index';
-//export * from './colorpicker/index';
+export * from './colorpicker/index';
 export * from './datepicker/index';
 export * from './dialog/index';
 export * from './menu/index';
